+++ arduino_due/exTask3/exTask3.cpp
@@ -20,7 +20,7 @@
 static PifTask *s_pstTask[3];
 
 
-static uint16_t _taskLedRedToggle(PifTask *pstTask)
+static uint32_t _taskLedRedToggle(PifTask *pstTask)
 {
 	static int nCount = 0, nToggle = 0;
 	static BOOL sw = LOW;
@@ -42,7 +42,7 @@ static uint16_t _taskLedRedToggle(PifTask *pstTask)
     return 0;
 }
 
-static uint16_t _taskLedYellowToggle(PifTask *pstTask)
+static uint32_t _taskLedYellowToggle(PifTask *pstTask)
 {
 	static int nCount = 0, nToggle = 0;
 	static BOOL sw = LOW;
@@ -64,7 +64,7 @@ static uint16_t _taskLedYellowToggle(PifTask *pstTask)
     return 0;
 }
 
-static uint16_t _taskLedGreenToggle(PifTask *pstTask)
+static uint32_t _taskLedGreenToggle(PifTask *pstTask)
 {
 	static int nCount = 0, nToggle = 0;
 	static BOOL sw = LOW;
