+++ arduino_nano/exTask1/exTask1.cpp
@@ -15,7 +15,7 @@
 #define TASK_SIZE				1
 
 
-static uint16_t _taskLedToggle(PifTask *pstTask)
+static uint32_t _taskLedToggle(PifTask *pstTask)
 {
 	static int nCount = 0;
 	static BOOL sw = LOW;
